pyproject.toml
[project]
name = "mkdocstrings-github"
version = "0.0.0"
description = "A GitHub Action handler for mkdocstrings"
authors = [
    { name = "Mark Hu", email = "watermarkhu@gmail.com" }
]
license =  { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10, <3.14"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Software Development",
    "Topic :: Software Development :: Documentation",
    "Topic :: Utilities",
    "Typing :: Typed",
]

dependencies = [
    "gitpython~=3.1.45",
    "mkdocstrings~=0.29",
    "mkdocs-autorefs~=1.4",
    "charset-normalizer~=3.4",
    "typing-extensions>=4.0; python_version < '3.11'",
]

[tool.uv.sources]
pygments = { git = "https://github.com/watermarkhu/pygments", branch = "matlab" }

[dependency-groups]
dev = [
    "ty~=0.0.1a19",
    "ruff~=0.12.9",
    "prek>=0.2.1",
    "pydantic~=2.10",
    "pytest~=8.4.1",
    "pytest-cov~=6.2.1",
    "beautifulsoup4>=4.13.3",
    "inline-snapshot~=0.27.2",
    "python-semantic-release~=10.3.0",
]
docs = [
    "mkdocs-material~=9.5",
    "markdown-exec~=1.10",
    "mkdocs-callouts~=1.15",
    "mkdocs-git-revision-date-localized-plugin~=1.3",
    "mkdocs-minify-plugin~=0.8",
    "mike~=2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
default-groups = ["dev", "docs"]
managed = true

[tool.ruff]
target-version = "py310"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",  # Python specific errors
    "W",  # Warnings
    "F",  # Pyflakes (code complexity)
    "I",  # isort
]
ignore = [
    "E501",  # Line too long
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.inline-snapshot]
storage-dir = "test/snapshots"
format-command = "ruff format --config pyproject.toml --stdin-filename {filename}"

[tool.ty.src]
include = ["src", "pyproject.toml"]
exclude = ["test"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mkdocstrings_handlers"]

[tool.semantic_release]
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "conventional"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"
version_toml = [
    "pyproject.toml:project.version",
]

[tool.semantic_release.branches.main]
match = "(?s).*"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.commit_parser_options]
patch_tags = ["fix", "perf", "chore"]

[tool.pytest.ini_options]
pythonpath = [
    "test"
]
markers = [
    "without_handler"
]
